feat: MainViewModel의 온보딩 로직 OnboardingViewModel로 이식
