+++ main-data/src/main/java/com/teamhy2/main/data/di/MainModule.kt
@@ -1,10 +1,8 @@
 package com.teamhy2.main.data.di
 
 import com.teamhy2.main.data.repository.DefaultStudyDayRepository
-import com.teamhy2.main.data.repository.DefaultWebViewRepository
 import com.teamhy2.main.data.repository.DefaultWiseSayingRepository
 import com.teamhy2.main.domain.repository.StudyDayRepository
-import com.teamhy2.main.domain.repository.WebViewRepository
 import com.teamhy2.main.domain.repository.WiseSayingRepository
 import dagger.Binds
 import dagger.Module
@@ -15,10 +13,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class MainModule {
-    @Binds
-    @Singleton
-    abstract fun bindWebViewRepository(impl: DefaultWebViewRepository): WebViewRepository
-
     @Binds
     @Singleton
     abstract fun bindWiseSayingRepository(impl: DefaultWiseSayingRepository): WiseSayingRepository
