+++ main-presentation/src/main/java/com/teamhy2/feature/main/MainViewModel.kt
@@ -1,8 +1,5 @@
 package com.teamhy2.feature.main
 
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.google.firebase.Firebase
@@ -14,9 +11,7 @@ import com.teamhy2.feature.main.model.MainUiState
 import com.teamhy2.hongikyeolgong2.notification.NotificationHandler
 import com.teamhy2.hongikyeolgong2.timer.prsentation.model.TimerUiModel
 import com.teamhy2.main.domain.repository.StudyDayRepository
-import com.teamhy2.main.domain.repository.WebViewRepository
 import com.teamhy2.main.domain.repository.WiseSayingRepository
-import com.teamhy2.onboarding.domain.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -30,47 +25,16 @@ import javax.inject.Inject
 class MainViewModel
     @Inject
     constructor(
-        private val webViewRepository: WebViewRepository,
         private val wiseSayingRepository: WiseSayingRepository,
-        private val userRepository: UserRepository,
         private val studyDayRepository: StudyDayRepository,
         val notificationHandler: NotificationHandler,
     ) : ViewModel() {
         private val _mainUiState = MutableStateFlow(MainUiState())
         val mainUiState: StateFlow<MainUiState> = _mainUiState.asStateFlow()
 
-        var urls by mutableStateOf<Map<String, String>>(emptyMap())
-            private set
-
-        val userUid: String by lazy {
-            Firebase.auth.currentUser?.uid ?: ""
-        }
-
-        private val _userExists: MutableStateFlow<Boolean> = MutableStateFlow(true)
-        val userExists: StateFlow<Boolean> = _userExists.asStateFlow()
-
         init {
-            getFirebaseUrls()
             getWiseSaying()
             getCalendarData()
-        }
-
-        fun checkUserExists(uid: String?) {
-            if (uid == null) {
-                _userExists.value = false
-                return
-            }
-
-            viewModelScope.launch {
-                _userExists.value = userRepository.checkUserExists(uid)
-            }
-        }
-
-        private fun getFirebaseUrls() {
-            viewModelScope.launch {
-                val urlList = webViewRepository.fetchFirebaseUrls()
-                urls = urlList.toMap()
-            }
         }
 
         private fun getWiseSaying() {
@@ -138,7 +102,7 @@ class MainViewModel
         }
 
         fun addStudyDay() {
-            val uid = userUid
+            val uid = Firebase.auth.currentUser?.uid ?: ""
             val startTime = mainUiState.value.startTime
 
             if (uid.isNotEmpty() && startTime != null) {
